Add ExtendModel mixin for extending an existing model with new traits.
